requirements.txt
streamlit
pandas
argostranslate
tqdm
python-pptx